Added dash to BLANK_ROW, corrected size of buffer for print_board, ternaried a couple if else blocks

+++ src/board.h
@@ -19,7 +19,7 @@ void set_board(board_t* board, int * values);
 void set_board_string(board_t* board, char* values);
 int set_board_file(board_t* board, char* filename);
 int get_cell_val(board_t* board, int row, int col);
-const char* print_row(board_t* board, char* buffer, int row);
+const char* print_row(board_t* board, char* buffer, uint8_t row);
 const char* print_blank_row();
 char* print_board(board_t* board);
 cell_t* get_row(board_t* board, cell_t* buffer, int row);
